Forgot to add files to index
Last commit:

Added a welcome activity to the project. This activity should be used to
display the app logo (GeoConcert) and maybe a loading bar or some short
description.

This activity can be used to fetch the GPS location and event list. Then
we can send the events directly to the eventlist activity by using an
intent.