4. As a user I want to view info about one artist. (2)
Status: In progress

File: EventInfoActivity
Refactored the code a lot. Removed unnecessarily complicated code and
changed single-line multi-line comments to use single line symbols.

File: ArtistInfoActivity
Added some logic to get the incoming Intent and fetch the artist string
so that this string could be sent to the FetchArtistAsyncTask object.