7. As a user I want to enter an artist and find events for similar artists. (3)
Status: In progress

File: FetchEventsForArtistAsyncTask
Added a class for fetching events for an artist. Extends the Async class
in the Android library. It stores the new events in the application
object.

File: SimilarArtistsActivity
A simple activity class that implements functionality for filling in a
list of artists and opening the ArtistInfoActivity if the user clicks an
artist in the list.

File: ArtistInfoActivity
Added a button for finding events for that artist called "Find Events".
It starts the FetchEventsForArtistAsyncTask...